Do not update last_event_time upon receiving a challenge
the problem is that we may not accept the challenge, but reseting will prevent us from sending challenges.

extreme example: setup bot to refuse bot challenges, but send a challenge every hour. the bot will never be able to send any challenge, because it will be spammed constantly by bot challenges (refused) that reset this last_event_time.