Fake Think
Pretend to think, like a human. Typically, weak levels of engines will move instantly, because they rely on drastic search limits to give humans a chance. For humans, this is annoying and unnatural to play against. And it can also be useful against bots, to reduce the speed at which instant moves bots will hit the lichess API.